Cleaned login and frontpage urls
Login link now uses /:locale/login instead of /sessions/new. In,
addition the unnecessary 'require_login' query string parameter
has been removed.

Changed a few redirects to frontpage to use frontpage_url (which
redirects to /:locale instead of /?locale=en) and also added a
an action to ApplicationController that redirects the actual root
('/') to the localized root based on I18n.locale value (which
may be the user's session locale or default value).